flask-api/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aniso8601==9.0.1
    # via flask-restful
attrs==23.2.0
    # via
    #   outcome
    #   trio
beautifulsoup4==4.12.3
    # via -r requirements.in
binance==0.3
    # via -r requirements.in
blinker==1.8.2
    # via flask
certifi==2024.6.2
    # via
    #   requests
    #   selenium
cffi==1.16.0
    # via trio
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   flask
    #   nltk
colorama==0.4.6
    # via
    #   click
    #   pytest
    #   tqdm
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
emoji==2.12.1
    # via -r requirements.in
exceptiongroup==1.2.1
    # via
    #   pytest
    #   trio
    #   trio-websocket
flask==3.0.3
    # via
    #   -r requirements.in
    #   flask-cors
    #   flask-restful
flask-cors==4.0.1
    # via -r requirements.in
flask-restful==0.3.10
    # via -r requirements.in
fonttools==4.53.0
    # via matplotlib
h11==0.14.0
    # via wsproto
idna==3.7
    # via
    #   requests
    #   trio
iniconfig==2.0.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
joblib==1.4.2
    # via nltk
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.9.0
    # via -r requirements.in
nltk==3.8.1
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   pandas
outcome==1.3.0.post0
    # via trio
packaging==24.1
    # via
    #   matplotlib
    #   pytest
pandas==2.2.2
    # via -r requirements.in
pillow==10.3.0
    # via matplotlib
pluggy==1.5.0
    # via pytest
pycparser==2.22
    # via cffi
pyparsing==3.1.2
    # via matplotlib
pysocks==1.7.1
    # via urllib3
pytest==8.2.2
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via
    #   flask-restful
    #   pandas
regex==2024.5.15
    # via nltk
requests==2.32.3
    # via -r requirements.in
selenium==4.22.0
    # via -r requirements.in
six==1.16.0
    # via
    #   flask-restful
    #   python-dateutil
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via trio
soupsieve==2.5
    # via beautifulsoup4
tomli==2.0.1
    # via pytest
tqdm==4.66.4
    # via nltk
trio==0.25.1
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.12.2
    # via
    #   emoji
    #   selenium
tzdata==2024.1
    # via pandas
urllib3[socks]==2.2.2
    # via
    #   requests
    #   selenium
websocket-client==1.8.0
    # via
    #   -r requirements.in
    #   selenium
werkzeug==3.0.3
    # via flask
wsproto==1.2.0
    # via trio-websocket
